Start testing Node map operator
- Move Node map to node.rs
